HUBSPOT_SYNC_PYTHON_PACKAGE/requirements.txt
# HUBSPOT_SYNC - Dependencias optimizadas SIN PANDAS
# Compatible con Python 3.13
python-dotenv==1.0.1
pyodbc==5.0.1
tabulate==0.9.0
requests==2.31.0
urllib3==2.2.2

# Nota: PANDAS ha sido removido para optimización
# El proyecto ahora usa estructuras de datos nativas de Python
# Resultado: Mayor velocidad y menor uso de memoria
# Actualizado para Python 3.13 - Versión más moderna y estable
